move bar chart core out of react component
